chore: clean up issue tests
Move all issue tests into issue_test.go and use common naming
convention.
